[flatland] Inject DisplayController into Renderer
Images used as textures need to have the buffer collections
they are associated with imported to the display so that
the display can recognize the images when we offline rendering
data to it.

This importing needs to happen at the same time as when we
set the Vulkan constraints on the renderer itself, so this CL
injects the display controller into the renderer so it can
control buffer collection management for the DC as well. This
also allows the DC buffer collection registration to happen
inside the flatland class without exposing the DC directly to
Flatland.

Lastly, this CL makes it so that the renderer and the display
share the same unique IDs for buffer collections, in order to
forego the need to keep track of mappings between them. To do
this, a global function has been added to display/utils.h to
generate ids atomically.

Testing - Unit tests in display_unittest.cc and engine_unittest.cc
updated to account for the change.

MULTIPLY: flatland_renderer_unittests (fuchsia): 30

Change-Id: I61290bfeb62b90663b637d562f7f868e449ff500
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/429458
Commit-Queue: Chris Fontas <cfontas@google.com>
Reviewed-by: Adam Gousetis <adamgousetis@google.com>
Reviewed-by: Evan Birenbaum <birenbaum@google.com>
Testability-Review: Evan Birenbaum <birenbaum@google.com>