[soundplayer] teach example util to play sounds from /tmp
This CL adds a feature to soundplayer_example that has it play all
the sounds in its isolated /tmp, if there are any such files. When
such files are found, the hard-coded sequence is not played.

This feature is useful for establishing that sounds will play
correctly with soundplayer.

Here's a host-side bash function that can be used to play a directory
full of files:

function playmanysounds() {
  if (($# != 1)); then
    echo "usage: playmanysounds <directory>"
    return 1
  fi
  fx shell mkdir /tmp/r/sys/fuchsia.com:soundplayer_example:0#meta:soundplayer_example.cmx
  for file in $1/*; do
    echo copying $file
    fx cp --to-target $file /tmp/r/sys/fuchsia.com:soundplayer_example:0#meta:soundplayer_example.cmx
  done
  fx shell run soundplayer_example
  fx shell rm -rf /tmp/r/sys/fuchsia.com:soundplayer_example:0#meta:soundplayer_example.cmx
}

Test: example...no automated tests
      The bash function above can be used for manual test.

Change-Id: Ie89ab5ea9fe9cd80a68874c129970ff15ad70235
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430350
Reviewed-by: Austin Foxley <afoxley@google.com>
Testability-Review: Austin Foxley <afoxley@google.com>