[iwlwifi] Enable driver but exclude from build
Instead of writing bind rules that always fail, just exclude the driver
from the build. Only the driver module itself is excluded, so the tests
that depend on the driver source will still be compiled and tested in
continuous integration. The only effect is that the driver manager will
not read the driver's info at runtime, or execute its bind rules, unless
the driver module is explicitly added to a build. And the driver won't
be compiled for most developers building a core or product build without
all the tests.

To run the driver, add the following to fx set:
--with-base //garnet/packages/prod:wifi_intel

Change-Id: I85bd26b84cccc8ecacdf7318330c6eabed2aebb9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430057
Commit-Queue: Alex Legg <alexlegg@google.com>
Reviewed-by: Louis Yung-chieh Lo <yjlou@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Testability-Review: Louis Yung-chieh Lo <yjlou@google.com>