[build] Add //build/config/zircon/standard.gni
Similar to //public/gn/config/standard.gni, and
accessible in both BUILD.zircon.gn and BUILD.gn files
as $zx_build_config/standard.gni, but only used
to hold the `standard_fuchsia_cpus` global list for
now.

This ensures that Zircon artifact toolchains are
always built for the right CPUs if this list is
updated in the future.

Bug: 54161
Testing: Should not change the build

Change-Id: Id527bfa7076ada01df783a4eb2d40ae591a4d5cc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430614
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: David Turner <digit@google.com>