[bootserver] Pass in authorized keys to Boot().
The bootserver tool takes in an authorized keys file as a flag, but I
incorrectly parsed that as a private key file because botanist used a
private key file and derived the authorized keys from it and I just
copied the logic. This change calls for the authorized_keys bytes to get
passed to Boot() instead of the private SSH signers so that the
bootserver tool can just pass its authorized_keys as is and botanist
will have to convert first before passing the keys to Boot().

Test: tested by CQ

Bug: 38517
Change-Id: I0a447846d392bb50cbaec800ef159960312333f5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430540
Commit-Queue: Ina Huh <ihuh@google.com>
Testability-Review: Ina Huh <ihuh@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>