[ui-input] refactor input-synthesis tests
Refactor the unit test code, to separate the collection
of events from the validation of those events. This enables
non-exact-comparison, which we'll need for a timestamp
test which we'll be adding shortly.

Along the way: remove the use of spawn(), in favor of
having the collection-and-validation-task run via
the run_singlethreaded() attribute macro.

Bug: 60438
Test: fx test input-synthesis-tests
Change-Id: I763df2abbcd663776a4a963ae9c45dd295a6f1e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430034
Commit-Queue: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Dragos Tiselice <dtiselice@google.com>
Testability-Review: Dragos Tiselice <dtiselice@google.com>