[fidl][rust] Add support for flexible enums
This CL adds support for flexible enums in the FIDL Rust bindings.
Unlike strict enums, they have an extra variant __Unknown which is
returned when the value on the wire is not recognized. Here is the API:

    from_primitive(prim: $prim_ty) -> Option<Self>  // same as strict
    into_primitive(self) -> $prim_ty                // same as strict
    unknown() -> Self                               // flexible only
    from_primitive_flexible(prim: $prim_ty) -> Self // flexible only
    validate(self) -> Result<Self, $prim_ty>        // flexible only
    is_unknown(&self) -> bool                       // flexible only

In addition, a flexible enum `Foo` gets a macro `FooUnknown!()` to be
used in match expressions. This expands to `_` normally, but in the
future there will be a build mode where it expands to the unknown
variant, allowing the detection of missing cases.

The __Unknown variant is marked deprecated to discourage using it
directly. If a user references Foo::__Unknown, they get this warning:

    WARNING: Use Foo::unknown() to construct, FooUnknown!() to exhaustively match

Test: fx test //src/lib/fidl/rust
Change-Id: Idee3f11a5655b6817908fa5affb2b036a0809e84
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428515
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Felix Zhu <fcz@google.com>
Testability-Review: Pascal Perez <pascallouis@google.com>
Testability-Review: Felix Zhu <fcz@google.com>