[drivers][time] Propagate clock.backstop boot arg to drivers
The RTC driver needs the same backstop as the rest of the system
in order for the kernel UTC time to stay in sync with the userspace
UTC clock. This is only necessary until UTC is removed from the kernel.

Bug:60452

Change-Id: Id60ba252d48a08ecd8ae5cc79dcf70e85050c9e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430794
Commit-Queue: Adam Lesinski <adamlesinski@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>