[bootserver] Only get images that exist.
When trying to get images with a local manifest, we assumed all images
existed and failed if it didn't. However, not all images are guaranteed
to exist and although we skip non-existent images from GCS, we didn't
have that check for the local images. This adds that check.

Change-Id: I43ea4168de13e34364f333fc5b3bce5090aff38f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430004
Commit-Queue: Ina Huh <ihuh@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Testability-Review: Ina Huh <ihuh@google.com>