[devices] Add first part of driver2 lib
This is part of the drivers-as-components work. Add enough of driver2
lib to support the implementation of driver_host2. In a future change,
more of driver2 lib will be added so that we can support an LLCPP-based
driver.

Bug: 33183
Change-Id: I004ab1dfd5616dd4f6d25c1df3aab738f91aaec5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430198
Commit-Queue: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>
Testability-Review: Abdulla Kamar <abdulla@google.com>