[build] Fix `fx format-code` and `fx lint` for FIDL files.
The 'fidl-format' and `fidl-lint` tools were recently moved to
the Fuchsia build, so update the scripts to look for the binaries in
the right place (and rebuild them if necessary).

For testing, do the following:

  - Modify a liberal .fidl file (e.g. messing with indentation).
  - Do "fx clean"
  - Run "fx format-code"
  - Run "fx lint"

Due to the clean build, this will rebuild the fidl-format and
fidl-lint tools respectively, then processing the .fidl file
with them. Then verify manually that it was fixed, since the
tools don't print anything in that case.

Bug: 49742
Testing: See above
Change-Id: I2bf2bff5e51561592b06c75c68f09426fd9d96de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430875
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: David Turner <digit@google.com>
Testability-Review: Shai Barack <shayba@google.com>