[cts] use prebuilt python to run test
Introduce python_host_test template to make this easy for other tests
to re-use. This results in tests that are totally self-contained, rather
than relying on python3.8 being in the PATH when they happen to get
executed. This should fix the attached issue, which is a result of tests
relying on the PATH.

Also add cts/build:tests to cts:cts so that the tests get run in CI,
which made it much easier to test this change.

Fixed: 60260
Change-Id: I6cca4b0b2329c89531f6a69e75c597098f0f91d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430344
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Reviewed-by: Jeremy Manson <jeremymanson@google.com>
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: John Shamoon <johnshamoon@google.com>
Commit-Queue: Gary Miguel <garymm@google.com>