[netstack] Observe v6 LL addr in if watcher test
Fix a flake in netstack_integration_test::test_interfaces_watcher
where the configuration of an IPv6 link-local address after the
interface is brought up may be observed and cause the assertion of
events to fail. After the interface is brought up, wait until the
interface up and the LL addresses are observed. After adding the IPv4
address, check that the new address set is equal to the LL addresses
observed previously plus the added address. The reverse of these
assertions are made when the IPv4 address is removed and when the
interface is brought down (which causes exactly one of the LL
addresses to be removed).

Includes some refactoring to improve code quality.

Test: fx test netstack_integration_test -- \
          --exact fidl::test_interfaces_watcher
Change-Id: I257dc313c73e20fa20c4cd88ebfec646459ef6d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/429887
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Testability-Review: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Tony Gong <gongt@google.com>