[modular][testing] Ensure test harness is torn down after test fixtures
TestHarnessFixture will now stop the test harness in TearDown, before
the fixture is destructed.

This ensures that agents and shells created in tests and stored in
fixture fields are destructed only after the harness is stopped.
In particular, this is needed to ensure that sessionmgr is connected
to the SessionShell protocol through its teardown process, as it may
need to call DetachView when deleting stories.

Bug: 60191
Change-Id: I5b5643222d8026eac0587377b98dec1baae3dc95
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430957
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Gabe Schine <thatguy@google.com>
Testability-Review: Gabe Schine <thatguy@google.com>