[libc] Replace strcmp call in start_main
There's a call to strcmp before we've finished initializing the
sanitizer runtime. However, this might crash, because strcmp might
itself be instrumented.

This isn't a problem for sanitizers like asan, because they fail open
when the shadow memory is zero initialized. But for tsan, the
instrumentation functions expect a TLS pointer variable to be
initialized, which causes a null pointer dereference crash in
__tsan_func_entry.

Bug: 45118
Change-Id: I82ba17048bf68ce3d1cb9f60329f672b595b0f8e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/429879
Commit-Queue: Matthew Dempsky <mdempsky@google.com>
Reviewed-by: Marco Vanotti <mvanotti@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>