[cts] Do CTS Dep Verification In a Python Script
Dependency verification is too much work to be done during generation
time, so it has to be done during build time. The Python script searches
for dependencies in the allowed_cts_deps list and looks for a CTS file
if the dependency isn't there. CTS files are written at generation time,
and are assumed to be correct until the build checks the dependencies.

Renames the cts_element targets because it's used as part of the output
of the inner action now.

Change-Id: I009dc47a077b23897efdfc8260b2f9dd52bf6eef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419484
Commit-Queue: John Shamoon <johnshamoon@google.com>
Reviewed-by: Jeremy Manson <jeremymanson@google.com>
Reviewed-by: Victor Chang <vfcc@google.com>
Reviewed-by: Jerry Belton <jcecil@google.com>
Testability-Review: Jeremy Manson <jeremymanson@google.com>