[fatfs] Fix cluster not zeroed bug.
In a different CL, I changed the fs_tests so the disk is filled with
non-zero data which highlighted this bug: any clusters allocated to
the root directory would not get zeroed.

Fatfs relies on File::is_dir() to determine whether or not to zero a
cluster when allocating a new one. Previously, is_dir() would return
false for the root directory, which would cause extra root directory
clusters to not be zeroed. Fix this so that the clusters are correctly
zeroed.

Change-Id: Ia72a16e23e6f4eb87c2d4d45b8840e37cdaec7e4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/431959
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Simon Shields <simonshields@google.com>
Testability-Review: Simon Shields <simonshields@google.com>