[sync-from-stem][jiri] Error on untracked branches needing rebase
jiri emits a warning, but doesn't abort if there's branches in
dependencies that have no upstream but aren't rebased. In this case,
print out our own warning along with the command to rebase, as it's
otherwise easy to not notice that some subtree is out of date.

If the branch has an upstream, or is already based upon the desired
revision, no error is output and the sync will proceed normally.

Unfortunately this has to be implemented by regex grubbing through
jiri's output.

Test: python3 -m doctest -v sync-from-stem.py
Fixed: 60889
Change-Id: I5a2dd37e9c7b57731ac6224b3ba2e887ce26fd22
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432275
Testability-Review: Adam Perry <adamperry@google.com>
Commit-Queue: Scott Graham <scottmg@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>