[build][Wconversion] Remove "Wno-conversion" public config from glm_workaround.
The public config "Wno-conversion" in glm_workaround has hidden configs
of targets that built upon "glm_workaround". In order to make
smaller changes to fix "Wconversion" issues, we need to remove
"Wno-conversion" from glm_workaround, and instead we add the config
to all the targets that directly or indirectly depends on it.

Change-Id: I562159a872d7f82648904df84ae7e8d2227fc213
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432436
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: Shai Barack <shayba@google.com>