[minfs] Relax the fsck check for purged inodes.
When Minfs grows, the new inodes are not zeroed, so the magic for inodes
will not necessarily be zero, so the existing fsck check for this will
fail.

This has started to appear now because FVM's slice size was recently
changed to be 32 KiB rather than the 1 MiB that it was previously and
when Minfs formats the volume it only starts with 1 slice.

Also:

  + Changed the slice size for the file system tests to 32 KiB.

  + Ram-disks for the file system tests are now filled with a non-zero
    value to prevent us from inadvertently depending on it being zeroed.
    This required changes to the ram-disk code to allow setting a block
    size when passing in a VMO.

  + Added a log message to FVM binding so we can see the slice size.

  + Changed the file-system tests to use stderr rather than syslog
    because it's easier to debug.

Bug: b/169352760
Change-Id: I00fd6188298ea9a7d8f0b4d87c652d7246d11859
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/431954
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>
API-Review: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Chris Suter <csuter@google.com>
Testability-Review: Brett Wilson <brettw@google.com>