[storage][extractor] Add extractor library 1/n
The library in its current state provides a way to write an image
file. Storage software can use this library to dump the data that is
relevant to them.

The library is in rust. It provides C bindings(in c-bindings directory)
that can be used in C++ code base.

This patch adds only extent part of the library.

Test: Ran host side unit tests

Change-Id: I807b3772a5396f39ad3aa1241e3d2b8b5969a504
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427715
Commit-Queue: Vikram Auradkar <auradkar@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Testability-Review: Chris Suter <csuter@google.com>