[codec] Add const VMO access API to InternalBuffer
This API will be used by the AML encoder to grab a handle to the sysmem
allocated VMO for passing to the canvas driver, and could be used by the
decoder for similar purposes if needed.

For now test where other InternalBuffers tests are.

Test: amlogic_decoder_integration_tests
Change-Id: Ic44eae0ec13bfb9c6a7ab90241e692df1731e912
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432556
Reviewed-by: Dustin Green <dustingreen@google.com>
Testability-Review: Austin Foxley <afoxley@google.com>
Commit-Queue: Austin Foxley <afoxley@google.com>