[tools][fuzzing] Gate undercoat E2E test on environment variable
This CL switches the gate for the undercoat end-to-end tests to be based
on the presence of the UNDERCOAT_E2E_TESTS environment variable instead
of the absence of the `-short` flag to `go test`, so as to unblock work
on `go test` workflows.

This is an interim solution only: ideally, as tracked in
fxbug.dev/47225, this test would run cleanly in all environments
including CI. However, since it involves launching qemu against a
fuzzer-variant build, it is currently only usable manually.

This CL adds a comment explaining how to run the test manually, and also
fixes a few places that have bit-rotted due to the lack of CI coverage.

Fixed: 58836
Test: cd tools; go test ./fuzz
Change-Id: Iff405345f02c4cd800b527d33e39e58e56076596
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/422113
Commit-Queue: Cameron Finucane <eep@google.com>
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>
Reviewed-by: Aaron Green <aarongreen@google.com>
Testability-Review: Aaron Green <aarongreen@google.com>