[uuid] Make //src/lib/uuid host-compatible
The original //src/lib/uuid called zx_cprng_draw to generate random
bytes for UUID, which is not availabe on the host. In this change, we
use std::random_device to generate the random bytes on the host.

Note that in this CL we also modified
//zircon/system/ulib/gpt/BUILD.gn
because it has to modified together with this change. Otherwise, either
`fx build` or `fx set` will fail.

Change-Id: Ia78a13436f8f212411c277f1a34271d58ffb3f21
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432452
Commit-Queue: Rundong Du <rundongdu@google.com>
Testability-Review: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>