[swd-e2e-tests] Add packages.json support to util
As part of fxb/50750, support for the legacy "packages" file format in
updates is being removed in favor of "packages.json".

This change adds support to the util package to parse the packages.json
format[1] and fuchsia-pkg URLs[2] through an API that returns data in
the same format as the existing "ParsePackageList" function. A future
commit will remove this latter API. This change also introduces tests
for both APIs, and in particular that the APIs are compatible in the
returned data.

Bug: 50750

[1] https://fuchsia.dev/fuchsia-src/concepts/software_delivery/update_pkg
[2] https://fuchsia.dev/fuchsia-src/concepts/packages/package_url

Change-Id: I052084d32eaba8061328d82e40439a056a913fb1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430711
Testability-Review: Zach Kirschenbaum <zkbaum@google.com>
Commit-Queue: Devon H. O'Dell <dhobsd@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>