[syslog] Don't stop streaming until context canceled
I'm not quite clear on when log_listener is supposed to exit, but we
definitely don't want to stop streaming syslogs if this happens. And it
does seem to happen sometimes - I wrote fxrev.dev/428815 to have
botanist fail if syslogging stops because of a log_listener successful
exit, and one of the tryjobs did fail because of that.

I also rewrote the syslog tests to eliminate shared state between tests,
which makes them easier to reason about and makes hard-to-debug failures
less common.

Finally, I renamed the syslog tests target from just "tests" to
"syslog_test" so that it can be referenced as "syslog_tests" in test
multipliers.

This should resolve the syslog truncation described in fxbug.dev/55301,
which happened because log_listener preemptively exited with a zero
retcode, causing the syslogger to quietly shut down. Now, we'll rerun
log_listener when that happens.

Multiply: syslog_tests: 1000
Fixed: 55301
Change-Id: Icce837d7b7148c8c99a032c6abfdb3b295339d3e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430705
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>
Testability-Review: Ina Huh <ihuh@google.com>