[build] Clarify the contract of binaries.json
The documentation wasn't clear about the fact that prebuilt binaries,
which are not produced by the build, should also be listed there
(and that in this case, their paths cannot be used as Ninja targets).

For context, see [1] and the comment that explains the contract in
more details (explaining why the CL was later reverted).

Bug: None
Testing: No change to the build.

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/431295

Change-Id: I765e06c48ae9c3f9c173a7b7ecf56ba4c231502f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/431298
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>