[build] rename universe_manifests.json
We should have a name that's more descriptive. In the original
name, it's unclear what kind of manifests these are. E.g. are they blob
manifests or package manifests? Are they paths or the actual contents of
the manifests?

The new name (all_package_manifest_paths.json) is more descriptive. The
file is now "exactly what it sounds like".

Bug: 60221
Change-Id: I6f7fe15031265f9bb22fbab949d1c88f64410463
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432177
Reviewed-by: James Tucker <raggi@google.com>
Commit-Queue: Zach Kirschenbaum <zkbaum@google.com>