[audio_core] Use sinc sampler if we might micro-SRC
To reconcile a client clock with a device clock, we might adjust
either of these (if they are adjustable); otherwise we insert a
micro-SRC frame-rate-conversion factor. Even if the device clock
supports fine-tuning, it may already be controlled by another
so we might use micro-SRC in that case as well. Only when the
client clock is adjustable ("flexible") are we guaranteed to NOT
need micro-SRC for synchronization.

Going forward, EXCEPT for the flexible clock scenario, we always
use the WindowedSinc resampler, as we may need its higher
fidelity when applying micro-SRC for clock synchronization.

Test: pending
Fixed: 60061

Change-Id: I5a3289746b0eb8ec60791c8173a39f268727e8c3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427561
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>
Testability-Review: Tom Bergan <tombergan@google.com>