[audio_core] Improved micro-SRC
This CL eliminates oscillation in the micro-SRC by quantizing the
adjustment rate so that it is a PPM measurement, just as the inputs
to the clock reconciliation process are.

In doing so, this CL also brings forward some of the reorganization
done in the coming "flexible" clock CL (for example, in PidControl).
As a result, the flexible CL will be just additions for that mode.

Clock tests using MixStage are split into their own unittest file.

Test: existing unittests reworked and tightened up

Change-Id: Ie28af5a049b3fdf3a4b2757a9194a5f78ee40c66
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/431080
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>
Testability-Review: Tom Bergan <tombergan@google.com>