[rust] Use a slice for channel write() and call() handles argument
Currently, the signature for channel write is:
pub fn write(&self, bytes: &[u8], handles: &mut Vec<Handle>)
  -> Result<(), Status>
However, slices are more idiomatic in rust for input arguments
because of deref coercion, which allows both Vec and non-Vec
types as input.

The reason Vec is used is to invalidate the Handle list
(with .set_len(0)) when handles are written to the channel,
however it is also possible to do this by directly setting
the items in the list to Handle::invalid(), which does not
require a Vec type.

However, this:
- changes the API to not resize the input
- has slightly more overhead in performance, though this seems
to be negligible and perhaps not worth optimizing if it isn't
a problem

Change-Id: Ia44efcb0428687b5d1e21052f5e440fdfa0188ae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/429815
Reviewed-by: Craig Tiller <ctiller@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Testability-Review: Craig Tiller <ctiller@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>