[build][alln] Migrate //zircon/system/ulib/zircon:zircon-headers to Fuchsia build.
This CL really allows building the Zircon VDSO (i.e. libzircon.so)
directly from the Fuchsia build. However, this is currently only used
to handle the .../zircon:zircon-headers target.

Migrating the .../zircon:zircon target entirely requires fixing the
sysroot setup to not depend on the Zircon legacy "zircon" target and
will require more work, for a future CL.

It is possible to build libzircon.so directly with:

  fx build user.vdso-arm64-shared/libzircon.so

The (stripped) result is byte-identical to the Zircon-produced library,
except for 8 build-id bytes. This was verified by comparing the binary
values of the two files, using the following command:

  diff -burN \
    <(xxd out/default.zircon/user.vdso-arm64-clang.shlib/obj/system/ulib/zircon/libzircon.so) \
    <(xxd out/default/user.vdso_arm64-shared/libzircon.so)

Bug: 58364
Testing: Should not change the build.
Change-Id: I253db14717b9778b634b4778c95a69ba63493f1d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432696
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: P.Y. Laligand <pylaligand@google.com>