[fdio] Enable retry of send/recv on connecting stream socket.
When the socket is still in connecting state, enable the callers of
stream socket read/write to retry the request after the state becomes
connected. This is as per RFC793, page 56, 58, for TCP send/recv.
In fdio, we have information about the socket state (connecting or
connected) and that can be used to handle this requirement.

Bug: 39041
Test: fx test -od netstack-bsdsocket-c-api-test -- \
	  --gtest_filter=NetStreamTest/BeforeConnectTest.BeforeConnect/*
Change-Id: I2f55c95ea066ba43f0fb330113388d078b9f245e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/429554
Commit-Queue: Mithun Iyer <iyerm@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Testability-Review: Tamir Duberstein <tamird@google.com>