[dart] Add dart tools to tool_paths.json
This will expose the paths of dart tools like blobstats to consumers of
the build, such as recipes. This is necessary in order for the recipes
to invoke blobstats using executable from the output directory, rather
than invoking blobstats' main.dart as a script using the prebuilt dart
interpreter. This will mean that recipes no longer need to know the path
to the prebuilt dart interpreter or to blobstats' main.dart source file.

Bug: 58254
Change-Id: I2b01798d0b3ca84733e3e2db7a6d861f06e8d19e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/433194
Reviewed-by: Chase Latta <chaselatta@google.com>
Reviewed-by: Gary Miguel <garymm@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>