[build][alln] Move //zircon/system/ulib/vdso-code-header to Fuchsia build.
This CL moves the vdso-code-header target completely to the Fuchsia
build, since it is no longer used by the Zircon one. Its purpose is
to expose the auto-generated vdso-code.h file (which contains macro
constants specific to the VDSO binary) to a few tests in the Fuchsia
build.

This requires adding a new empty/fake kernel-specific zircon toolchain
suite, which isn't used to build anything, except redirect some
targets. A full definition for the kernel toolchain will happen in
future CLs.

This also removes the last remaining entry from
legacy_targets-$cpu.json. A future CL will remove the generation
of that file entirely.

For testing, the content of the two following files are compared:

  out/default/kernel_x64/gen/zircon/kernel/lib/userabi/vdso-code.h
  out/default.zircon/kernel-x64-clang/gen/kernel/lib/userabi/vdso-code.h

And the only differences are in the values for the VDSO_FILENAME
and VDSO_BUILD_ID_NOTE_BYTES, which are both expected. All other
macro values are exactly the same.

Bug: 59926
Bug: 58463
Testing: See above.

Change-Id: I1e785982422c9320ac56c24a5a86c88bd46545da
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432697
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: P.Y. Laligand <pylaligand@google.com>