[tools] Add blobstats to //tools:tools target
We'd like to have recipes start consuming the blobstats executable
produced by the build, rather than executing blobstats as a script, so
we need to hook it into the build graph.

Once this transition is complete, we can move the blobstats source code
from //scripts (which is supposed to contain code that's only run by fx)
into //tools, which is a more standard location for code that's also run
by infra.

Also, specify $host_toolchain for other dart tools, since they are only
intended to be run on the host.

Bug: 58254
Change-Id: I58d8badf290077743935106ff343283482cbff02
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/433326
Reviewed-by: Gary Miguel <garymm@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>