[storage][extractor] Add extractor library 2/n
The library in its current state provides a way to write an image
file. Storage software can use this library to dump the data that is
relevant to them.

The library is in rust. It provides C bindings(in c-bindings directory)
that can be used in C++ code base.

This patch adds only extent cluster part of the library.

Test: Ran host side unit tests

Change-Id: I830a366399acedf3291a84cf017604950f3c17ef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432634
Commit-Queue: Vikram Auradkar <auradkar@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Testability-Review: Chris Suter <csuter@google.com>