[fx][formatting] Select and build the right doc-checker when formatting
This fixes a bug in which the `fx format-code` was not consulting
metadata to figure out which doc-checker to use (it hard-coded a path
under out/default). We update the formatting script to rely on
tool_paths metadata to figure out where to find the tool.

We also superficially update the fidl-format logic as it is inheritently
similarly and so it would be nicer to have a consistency among the
names/spellings/etc. across the two cases.

Change-Id: Id7e957830de159df8c2cae44c0751324f34f5dc6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/433974
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Testability-Review: Renato Mangini Dias <mangini@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>