[timekeeper] Add a clock manager and estimator.
The estimator is a placeholder within which the kalman filter and
frequency tracker can live in the future (without requiring any async
code!!). The clock manager consolidates some more of the logic
previously in main and will be the site of the clock update logic once
we start to use slewing. As of this CL there should be no change in
behavior of the clock or of the diagnostics.

Bug: 56858

Change-Id: I7b848f009cff29cac100f4286aae30fe2dcdd0e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432625
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Testability-Review: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>