[scrutiny] Add Fuchsia Volume Manager Extractor.
This change adds a new FvmParser to the scrutiny_utils crate which
enables parsing a non-spare Fuchsia Volume and extracting the partitions
from it. It does this by creating virtual slice -> physical slice
mappings for each partition and reading that back in order directly into
a buffer for each partition. The type is also provided in the definition
along with the buffer for the partition.

This isn't designed to scale to large disks as the intention is that this is
used only for StorageRamdisks inside ZBIs. This is a step towards parsing out
the BlobFS inode based filesystem within the BlobFS partition which exists
inside the FVM.

This change also introduces a new command tool.fvm.extract which allows
extracting a FVM into its component filesystems. This command will not
attempt to extract or interpret the internal file systems.

Usage:
tool.fvm.extract --input fvm.blk --output /tmp/fvm

Output
/tmp/fvm/blobfs.blk
/tmp/fvm/minfs.blk

Change-Id: I92869f5ca757c3b280ddfe921441d9a414b1aed6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/422156
Testability-Review: Benjamin Wright <benwright@google.com>
Reviewed-by: Yu Shan <shanyu@google.com>
Reviewed-by: Daniel Hwang <arkay@google.com>
Commit-Queue: Benjamin Wright <benwright@google.com>