[devmgr_launcher] Remove access to /boot and /system
All tests which depend on this library should be packaged and avoid
access to /boot and /system in order to be hermetic. As a result we can
assume access to /pkg and simplify our library interface by not
requiring packaged test to specify path_prefix.

This change does break the ability to write an isolated devmgr based
test in a bringup build for now, however at some point in the future,
this capability may be restored if we add support for a component based
test runner in bootfs.

Change-Id: Iaf60644778528cdd3c6905fe3fc4913133794298
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409220
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Suraj Malhotra <surajmalhotra@google.com>
Testability-Review: Jocelyn Dang <jocelyndang@google.com>