[time][flake] Fix timekeeper integ test flake
To avoid faking out the entirety of the network state service, the
timekeeper integration test injects a fake that does only enough to
allow the setup of the network check to succeed, but not the entire
check. This is okay because timekeeper tries to sync time anyway even
if the network check fails, so long as the setup succeeds.
However, the setup involves making a request through the connection
made to the network service. If the connection is broken before that
request is made, timekeeper crashes and the test fails.
This change ensures that the channel held by the fake is held until
the required request is made.

Multiply: timekeeper-integration: 30

Change-Id: Idc6e660517492c52d8908d349d6c25f1eea27d6c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434044
Reviewed-by: Jody Sankey <jsankey@google.com>
Testability-Review: Jody Sankey <jsankey@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>