[tools] Move blobstats to infratools
blobstats is used by the infrastructure, so it must be included in all
builders (//bundles:infratools is built by all builders). It always runs
on the same machine as the build so it only needs to be built for
$host_toolchain.

Also clean up generation of toolchain-specific tools, and stop
building x64 and arm64 versions for some tools that only need to run on
the builder: artifactory, testsharder, and size_checker.

Bug: 58254
Change-Id: Id32f55de97af8162872eb1c746d9e0ea30f2a53b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/433814
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Gary Miguel <garymm@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>