[testsharder] Handle the case where test count < max-shards-per-env.
The broken builder in the attached bug is failing to create shards
because it's missing test durations so it defaults to the
maxShardsPerEnvironment for the number of shards to create. However, the
number of tests it runs is less than that, so it gets an indexing error
when trying to access the tests in a shard because some of the shards
ends up with no tests.

Bug: 61212
Change-Id: I02ea91eb30e5f363e8c37f1acf35972cebdb765f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434313
Reviewed-by: Oliver Newman <olivernewman@google.com>
Testability-Review: Oliver Newman <olivernewman@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>