[debugger] Use maximum module extent.
Compute the extent of ELF binaries and don't match symbols outside of
these bounds. Without this, random heap addresses can match the last
symbol in the nearest mapped shared library.

Add a constant limit on the distance away from ELF symbols we can match
addresses as a backstop for cases where this fails. The goal is to find
a symbol if it's there but not to be too confusing.

Updates the module symbol test initialization to assert that
initialization works. If some files are missing the previous code would
crash rather than report an error.

Bug: 42243
Change-Id: I97ef8be3e9848c49aeeeb5252fa89cfae329fc4d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434474
Reviewed-by: Casey Dahlin <sadmac@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>